+++ lib/rspec/mocks/stub_const.rb
@@ -4,15 +4,54 @@ module RSpec
     # constant stubbing.
     # @api private
     module RecursiveConstMethods
+      # We only want to consider constants that are defined directly on a
+      # particular module, and not include top-level/inherited constants.
+      # Unfortunately, the constant API changed between 1.8 and 1.9, so
+      # we need to conditionally define methods to ignore the top-level/inherited
+      # constants.
+      #
+      # Given `class A; end`:
+      #
+      # On 1.8:
+      #   - A.const_get("Hash") # => ::Hash
+      #   - A.const_defined?("Hash") # => false
+      #   - Neither method accepts the extra `inherit` argument
+      # On 1.9:
+      #   - A.const_get("Hash") # => ::Hash
+      #   - A.const_defined?("Hash") # => true
+      #   - A.const_get("Hash", false) # => raises NameError
+      #   - A.const_defined?("Hash", false) # => false
+      if Module.method(:const_defined?).arity == 1
+        def const_defined_on?(mod, const_name)
+          mod.const_defined?(const_name)
+        end
+
+        def get_const_defined_on(mod, const_name)
+          if const_defined_on?(mod, const_name)
+            return mod.const_get(const_name)
+          end
+
+          raise NameError, "uninitialized constant #{mod.name}::#{const_name}"
+        end
+      else
+        def const_defined_on?(mod, const_name)
+          mod.const_defined?(const_name, false)
+        end
+
+        def get_const_defined_on(mod, const_name)
+          mod.const_get(const_name, false)
+        end
+      end
+
       def recursive_const_get(const_name)
-        const_name.split('::').inject(Object) { |mod, name| mod.const_get name }
+        const_name.split('::').inject(Object) { |mod, name| get_const_defined_on(mod, name) }
       end
 
       def recursive_const_defined?(const_name)
         const_name.split('::').inject([Object, '']) do |(mod, full_name), name|
           yield(full_name, name) if block_given? && !mod.is_a?(Module)
-          return false unless mod.const_defined?(name)
-          [mod.const_get(name), [mod, name].join('::')]
+          return false unless const_defined_on?(mod, name)
+          [get_const_defined_on(mod, name), [mod, name].join('::')]
         end
       end
     end
@@ -139,7 +178,7 @@ module RSpec
       class DefinedConstantReplacer < BaseStubber
         def stub
           @context = recursive_const_get(@context_parts.join('::'))
-          @original_value = @context.const_get(@const_name)
+          @original_value = get_const_defined_on(@context, @const_name)
 
           constants_to_transfer = verify_constants_to_transfer!
 
@@ -160,7 +199,7 @@ module RSpec
 
         def transfer_nested_constants(constants)
           constants.each do |const|
-            @stubbed_value.const_set(const, original_value.const_get(const))
+            @stubbed_value.const_set(const, get_const_defined_on(original_value, const))
           end
         end
 
@@ -202,9 +241,9 @@ module RSpec
         def stub
           remaining_parts = @context_parts.dup
           @deepest_defined_const = @context_parts.inject(Object) do |klass, name|
-            break klass unless klass.const_defined?(name)
+            break klass unless const_defined_on?(klass, name)
             remaining_parts.shift
-            klass.const_get(name)
+            get_const_defined_on(klass, name)
           end
 
           context = remaining_parts.inject(@deepest_defined_const) do |klass, name|