+++ Changelog.md
@@ -6,6 +6,9 @@ Bug fixes
 * Don't modify `dup` on classes that don't support `dup` (David Chelimsky)
 * Fix `any_instance` so that it works properly with methods defined on
   a superclass. (Daniel Eguzkiza)
+* Fix `stub_const` so that it works properly for nested constants that
+  share a name with a top-level constant (e.g. "MyGem::Hash"). (Myron
+  Marston)
 
 ### 2.11.1 / 2012-07-09
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.11.0...v2.11.1)